feat:매칭 기능 구현 #BACK-233 #Resolves: #232, #233

+++ backend/src/main/java/signiel/heartsigniel/model/member/dto/MemberInfo.java
@@ -10,7 +10,7 @@ public class MemberInfo {
 
     private Long memberId;
     private String nickname;
-    private int rating;
+    private Long rating;
     private String gender;
     private String job;
     private LocalDate birth;
